backend/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==3.2.16
django-extra-fields==3.0.2
django-filter==23.5
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==20.1.0
idna==3.6
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
oauthlib==3.2.2
Pillow==9.5.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.4.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.7
zipp==3.15.0
